+++ vmworkstation/config.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/terraform/helper/schema"
 )
 
+// Config this struct is to storage the data of configuration to the web client
 type Config struct {
 	User         string
 	Password     string
@@ -16,6 +17,7 @@ type Config struct {
 	Debug        bool
 }
 
+// NewConfig_bak this method is usefull tu create a new config for the client without parameters
 func NewConfig(d *schema.ResourceData) (*Config, error) {
 	if d.Get("user").(string) == "" || d.Get("password").(string) == "" || d.Get("url").(string) == "" {
 		err := fmt.Errorf("User, Password and URL that required parameters")
@@ -32,6 +34,7 @@ func NewConfig(d *schema.ResourceData) (*Config, error) {
 	return config, nil
 }
 
+// Client_bak this method give a new client to communicate with the API REST to VmWare Workstation
 func (co *Config) Client() (*wsapiclient.Client, error) {
 	client, err := wsapiclient.New()
 	// log.Printf("[VMWS] Fi: config.go Fu: Client Obj: client after %#v\n", client)