+++ CHANGELOG.md
@@ -5,3 +5,17 @@
 NOTES:
 
 * Was generate this documentation and the first structure of folders for the provider module.
+
+## 0.1.9 (January 05, 2021)
+
+NOTES:
+
+* We add some functions and refactoring
+* We change the method to use with the API VmWare Version 16.02
+
+## 0.2.0 (April 18, 2021)
+
+NOTES:
+
+* We have solve some issues
+* Also, we have added a new foncionality to register the VM in the gui of VmWare